rename: realtime데이터를 불러와 정렬하는 기능을 상위 컴포넌트로 이동
