+++ src/api/supabaseSSR/supabase.ts
@@ -1,13 +1,13 @@
 import { createClient } from '@/utils/supabase/server';
 
-const supabase = createClient();
-
 /*
  * SSR
  */
 
 // my-owl 페이지의 유저 프로필 정보에 필요한 Data를 supabase, supabase auth에서 가져와 반환
 export const getUserProfileData = async () => {
+  const supabase = await createClient();
+
   const {
     data: { user },
     error
