+++ src/utils/supabase/userDataFetch.ts
@@ -1,9 +1,8 @@
 'use client';
 
 import { getCurrentUserData, getRoomUsersData } from '@/api/supabaseCSR/supabase';
-import { RoomData } from '@/components/room/sidebar/user/UserList';
 
-export const userDataFetch = async (id: string, setRoomData: (roomData: RoomData) => void) => {
+export const userDataFetch = async (id: string) => {
   const { user } = await getCurrentUserData();
   const userRoomData = await getRoomUsersData(id);
   const currentUserId = user.id;
@@ -13,5 +12,6 @@ export const userDataFetch = async (id: string, setRoomData: (roomData: RoomData
   const otherUsers = userRoomData.filter((user) => !user.is_admin && user.user_id !== currentUserId);
 
   const sortedUsers = [...adminUser, ...currentUser, ...otherUsers];
-  setRoomData(sortedUsers);
+
+  return sortedUsers;
 };
